refactor(skill): align hook cleanup test ownership
